requirements.txt
asgiref==3.5.2
backports.zoneinfo;python_version<"3.9"
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
pytz==2022.6
sqlparse==0.4.3
djangorestframework-simplejwt==5.3.1
